// contracts/typescript/models/NotificationView.ts
/* tslint:disable */
/* eslint-disable */
/**
 * HomeOfficeReservation API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues, parseDate, parseDateTime, serializeDate, serializeDateTime } from '../runtime';
import type { NotificationDestination } from './NotificationDestination';
import {
    NotificationDestinationFromJSON,
    NotificationDestinationFromJSONTyped,
    NotificationDestinationToJSON,
    NotificationDestinationToJSONTyped,
} from './NotificationDestination';

/**
 *
 * @export
 * @interface NotificationView
 */
export interface NotificationView {
    /**
     *
     */
    createdAt: Date;
    /**
     *
     */
    destination: NotificationDestination | null;
    /**
     *
     */
    eventType: string;
    /**
     *
     */
    historical: boolean;
    /**
     *
     */
    id: string;
    /**
     *
     */
    readAt: Date | null;
}

/**
 * Check if a given object implements the NotificationView interface.
 */
export function instanceOfNotificationView(value: object): value is NotificationView {
    if (!('createdAt' in value) || value['createdAt'] === undefined) return false;
    if (!('destination' in value) || value['destination'] === undefined) return false;
    if (!('eventType' in value) || value['eventType'] === undefined) return false;
    if (!('historical' in value) || value['historical'] === undefined) return false;
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('readAt' in value) || value['readAt'] === undefined) return false;
    return true;
}

export function NotificationViewFromJSON(json: any): NotificationView {
    return NotificationViewFromJSONTyped(json, false);
}

export function NotificationViewFromJSONTyped(json: any, ignoreDiscriminator: boolean): NotificationView {
    if (json == null) {
        return json;
    }
    return {

        'createdAt': (json['createdAt'] == null ? json['createdAt'] : parseDateTime(json['createdAt'])),
        'destination': NotificationDestinationFromJSON(json['destination']),
        'eventType': json['eventType'],
        'historical': json['historical'],
        'id': json['id'],
        'readAt': (json['readAt'] == null ? null : parseDateTime(json['readAt'])),
    };
}

export function NotificationViewToJSON(json: any): NotificationView {
    return NotificationViewToJSONTyped(json, false);
}

export function NotificationViewToJSONTyped(value?: NotificationView | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {

        'createdAt': value['createdAt'] == null ? value['createdAt'] : serializeDateTime(value['createdAt']),
        'destination': NotificationDestinationToJSON(value['destination']),
        'eventType': value['eventType'],
        'historical': value['historical'],
        'id': value['id'],
        'readAt': value['readAt'] == null ? value['readAt'] : serializeDateTime(value['readAt']),
    };
}